requirements.txt
transformers
datasets
torch
python-docx
scikit-learn
optuna
trl
evaluate